Fixed pfp upload and get along with account UI
Originally, pfps are uploaded to some upload directory that assumed that the computer uploaded in the root directory or something like that (and that it's running on Windows). Now, it's fixed with a ternary operator checking that it's either in the old or the new. The UI is changed in the login/signup/change account sections to look more uniform with the rest of the interface.